[yocto] Version an environment for building the litecell firmware
The vm performs so much disk access that it will quickly saturate the
ram buffer, using all provided memory. This will lead to slow
performance on a host machine with only 16GB of ram total that is
being used for other work at the same time. The compilation appears
compute bound on 3 cores with 8GB, so increasing the memory even more
does not greatly improve performance.